+++ R/utilities-scenario-configuration.R
@@ -36,7 +36,7 @@ readScenarioConfigurationFromExcel <- function(scenarioNames = NULL, projectConf
   # Define the casting functions to cast columns to specific type
   colTypes <- c(
     "text", "text", "text", "logical", "text",
-    "text", "numeric", "text", "logical",
+    "text", "text", "text", "logical",
     "numeric", "text", "text", "text"
   )
 
@@ -112,12 +112,11 @@ readScenarioConfigurationFromExcel <- function(scenarioNames = NULL, projectConf
     }
 
     # Simulation time
-    simTime <- data$SimulationTime
-    simTimeUnit <- data$SimulationTimeUnit
     # Set the time only if new value is defined
-    if (!is.na(simTime)) {
-      scenarioConfiguration$simulationTime <-
-        ospsuite::toBaseUnit(ospDimensions$Time, values = simTime, unit = simTimeUnit)
+    if (!is.na(data$SimulationTime)) {
+      scenarioConfiguration$simulationTime <- data$SimulationTime
+
+      scenarioConfiguration$simulationTimeUnit <- data$SimulationTimeUnit
     }
 
     # Individual id
@@ -220,4 +219,39 @@ setApplications <- function(simulation, scenarioConfiguration) {
       stop(messages$noPopulationIdForPopulationScenario(scenarioConfiguration$scenarioName))
     }
   }
+}
+
+# Parse simulation time intervals
+.parseSimulationTimeIntervals <- function(simulationTimeIntervalsString) {
+  # Check if the simulation time intervals are defined
+  if (is.null(simulationTimeIntervalsString)) {
+    return(NULL)
+  }
+
+  # Split the string by ';'
+  simulationTimeIntervals <- strsplit(x = simulationTimeIntervalsString, split = ";", fixed = TRUE)[[1]]
+  # Split each interval by ','
+  simulationTimeIntervals <- strsplit(x = simulationTimeIntervals, split = ",", fixed = TRUE)
+  # Convert to numeric
+  simulationTimeIntervals <- lapply(simulationTimeIntervals, as.numeric)
+  # Validate that all are numeric
+  validateIsNumeric(simulationTimeIntervals)
+  # Validate that all are positive
+  if (any(unlist(simulationTimeIntervals) < 0)) {
+    stop(messages$stopWrongTimeIntervalString(simulationTimeIntervalsString))
+  }
+  # Validate all intervals are of length 3
+  if (any(sapply(simulationTimeIntervals, length) != 3)) {
+    stop(messages$stopWrongTimeIntervalString(simulationTimeIntervalsString))
+  }
+  # Validate all resolution entries are greater than 0
+  if (any(sapply(simulationTimeIntervals, function(x) x[3] <= 0))) {
+    stop(messages$stopWrongTimeIntervalString(simulationTimeIntervalsString))
+  }
+  # Validate all start values are smaller than end values
+  if (any(sapply(simulationTimeIntervals, function(x) x[1] >= x[2]))) {
+    stop(messages$stopWrongTimeIntervalString(simulationTimeIntervalsString))
+  }
+
+  return(simulationTimeIntervals)
 }