+++ tests/testthat/test-utilities-scenario-configuration.R
@@ -42,16 +42,17 @@ test_that("It creates the correct scenario", {
   expect_equal(scenarioConfigurations[[scenarioNames]]$individualId, "Indiv1")
   expect_equal(scenarioConfigurations[[scenarioNames]]$modelFile, "Aciclovir.pkml")
   expect_equal(scenarioConfigurations[[scenarioNames]]$paramSheets, enum(enumValues = "Global"))
-  expect_equal(scenarioConfigurations[[scenarioNames]]$pointsPerMinute, 1)
   expect_equal(scenarioConfigurations[[scenarioNames]]$scenarioName, "TestScenario")
   expect_equal(scenarioConfigurations[[scenarioNames]]$setTestParameters, FALSE)
   expect_equal(scenarioConfigurations[[scenarioNames]]$simulateSteadyState, FALSE)
-  expect_equal(scenarioConfigurations[[scenarioNames]]$simulationTime, toUnit(
-    quantityOrDimension = ospDimensions$Time,
-    values = 24,
-    targetUnit = ospUnits$Time$min,
-    sourceUnit = ospUnits$Time$h
-  ))
+  expect_equal(
+    scenarioConfigurations[[scenarioNames]]$simulationTime,
+    list(c(0, 24, 60))
+  )
+  expect_equal(
+    scenarioConfigurations[[scenarioNames]]$simulationTimeUnit,
+    ospUnits$Time$h
+  )
   expect_equal(scenarioConfigurations[[scenarioNames]]$simulationRunOptions, NULL)
   expect_equal(scenarioConfigurations[[scenarioNames]]$simulationType, "Individual")
   expect_equal(scenarioConfigurations[[scenarioNames]]$steadyStateTime, 1000)
@@ -69,16 +70,17 @@ test_that("It creates multiple correct scenarios", {
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$individualId, "Indiv1")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$modelFile, "Aciclovir.pkml")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$paramSheets, enum(enumValues = "Global"))
-  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$pointsPerMinute, 1)
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$scenarioName, "TestScenario")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$setTestParameters, FALSE)
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulateSteadyState, FALSE)
-  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationTime, toUnit(
-    quantityOrDimension = ospDimensions$Time,
-    values = 24,
-    targetUnit = ospUnits$Time$min,
-    sourceUnit = ospUnits$Time$h
-  ))
+  expect_equal(
+    scenarioConfigurations[[scenarioNames[[1]]]]$simulationTime,
+    list(c(0, 24, 60))
+  )
+  expect_equal(
+    scenarioConfigurations[[scenarioNames[[1]]]]$simulationTimeUnit,
+    ospUnits$Time$h
+  )
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationRunOptions, NULL)
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationType, "Individual")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$steadyStateTime, 1000)
@@ -88,16 +90,17 @@ test_that("It creates multiple correct scenarios", {
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$individualId, "Indiv")
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$modelFile, "Aciclovir.pkml")
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$paramSheets, enum(enumValues = "Global"))
-  expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$pointsPerMinute, 1)
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$scenarioName, "TestScenario2")
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$setTestParameters, FALSE)
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$simulateSteadyState, TRUE)
-  expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$simulationTime, toUnit(
-    quantityOrDimension = ospDimensions$Time,
-    values = 12,
-    targetUnit = ospUnits$Time$min,
-    sourceUnit = ospUnits$Time$h
-  ))
+  expect_equal(
+    scenarioConfigurations[[scenarioNames[[2]]]]$simulationTime,
+    list(c(0, 1, 60), c(1, 12, 20))
+  )
+  expect_equal(
+    scenarioConfigurations[[scenarioNames[[2]]]]$simulationTimeUnit,
+    ospUnits$Time$h
+  )
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$simulationRunOptions, NULL)
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$simulationType, "Individual")
   expect_equal(scenarioConfigurations[[scenarioNames[[2]]]]$steadyStateTime, 500)
@@ -139,16 +142,17 @@ test_that("It creates a population scenario", {
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$populationId, "TestPopulation")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$modelFile, "Aciclovir.pkml")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$paramSheets, enum(enumValues = "Global"))
-  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$pointsPerMinute, 1)
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$scenarioName, "PopulationScenario")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$setTestParameters, FALSE)
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulateSteadyState, FALSE)
-  expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationTime, toUnit(
-    quantityOrDimension = ospDimensions$Time,
-    values = 12,
-    targetUnit = ospUnits$Time$min,
-    sourceUnit = ospUnits$Time$h
-  ))
+  expect_equal(
+    scenarioConfigurations[[scenarioNames[[1]]]]$simulationTime,
+    list(c(0, 12, 20))
+  )
+  expect_equal(
+    scenarioConfigurations[[scenarioNames[[1]]]]$simulationTimeUnit,
+    ospUnits$Time$h
+  )
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationRunOptions, NULL)
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$simulationType, "Population")
   expect_equal(scenarioConfigurations[[scenarioNames[[1]]]]$steadyStateTime, 1000)
@@ -256,5 +260,68 @@ test_that("It throws an error when reading wrong file structure for scenario con
         ), expectedColNames = expectedColumns), fixed = TRUE
       )
     }
+  )
+})
+
+
+# Test .parseSimulationTimeIntervals()
+test_that("It parses time intervals correctly", {
+  timeIntervalStringNull <- NULL
+  timeIntervalStringValidOneInterval <- "0, 1, 1"
+  timeIntervalStringValidTwoIntervals <- "0, 1, 1; 1, 2, 1"
+  timeIntervalStringStartAfterEnd <- "2, 1, 1"
+  timeIntervalStringResolutionZero <- "1, 2, 0"
+  timeIntervalStringInvaldiNonNumeric <- "1, 2, a"
+  timeIntervalStringNegative <- "-1, 1, 1"
+  timeIntervalStringInvalid <- "0, 1"
+  timeIntervalStringInvalidMultiple <- "0; 1; 1, 1; 2; 1"
+
+
+  expect_equal(
+    .parseSimulationTimeIntervals(timeIntervalStringNull),
+    NULL
+  )
+
+  expect_equal(
+    .parseSimulationTimeIntervals(timeIntervalStringValidOneInterval),
+    list(c(0, 1, 1))
+  )
+
+  expect_equal(
+    .parseSimulationTimeIntervals(timeIntervalStringValidTwoIntervals),
+    list(
+      c(0, 1, 1),
+      c(1, 2, 1)
+    )
+  )
+
+  expect_error(
+    .parseSimulationTimeIntervals(timeIntervalStringStartAfterEnd),
+    regexp = messages$stopWrongTimeIntervalString(timeIntervalStringStartAfterEnd),
+    fixed = TRUE
+  )
+
+  expect_error(
+    .parseSimulationTimeIntervals(timeIntervalStringResolutionZero),
+    messages$stopWrongTimeIntervalString(timeIntervalStringResolutionZero),
+    fixed = TRUE
+  )
+
+  expect_error(
+    .parseSimulationTimeIntervals(timeIntervalStringNegative),
+    messages$stopWrongTimeIntervalString(timeIntervalStringNegative),
+    fixed = TRUE
+  )
+
+  expect_error(
+    .parseSimulationTimeIntervals(timeIntervalStringInvalidMultiple),
+    messages$stopWrongTimeIntervalString(timeIntervalStringInvalidMultiple),
+    fixed = TRUE
+  )
+
+  expect_error(
+    .parseSimulationTimeIntervals(timeIntervalStringInvalid),
+    messages$stopWrongTimeIntervalString(timeIntervalStringInvalid),
+    fixed = TRUE
   )
 })